requirement.txt
streamlit
PyMuPDF
speechrecognition
pyttsx3
pyaudio
groq
langchain
langchain
langgraph
python-dotenv
langchain-groq
langchain-openai
langchain-google
langchain_core
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
PyMuPDF
fpdf
wikipedia
arxiv
utils